PDFBOX-1939: Store all stroke information in the graphics state
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1571581 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/PDExtendedGraphicsState.java
@@ -89,7 +89,7 @@ public class PDExtendedGraphicsState implements COSObjectable
         {
             if( key.equals( COSName.LW ) )
             {
-                gs.setLineWidth( getLineWidth().doubleValue() );
+                gs.setLineWidth( getLineWidth() );
             }
             else if( key.equals( COSName.LC ) )
             {
@@ -101,7 +101,7 @@ public class PDExtendedGraphicsState implements COSObjectable
             }
             else if( key.equals( COSName.ML ) )
             {
-                gs.setMiterLimit( getMiterLimit().doubleValue() );
+                gs.setMiterLimit( getMiterLimit() );
             }
             else if( key.equals( COSName.D ) )
             {
